requirements.txt
flask
python-dotenv
azure-ai-textanalytics
pypdf
python-docx
